preliminary pi estimation routine implemented

+++ src/linear_distance_pi/mod.rs
@@ -8,11 +8,11 @@ pub struct Args {
 
   /// Number of iterations to run
 	#[arg(short, long, default_value_t = 10)]
-  pub iter: usize,
+  pub iter: u64,
 
 	/// Power of two used to get portion of the unit circle (2^n)
 	#[arg(short, long, default_value_t = 4)]
-  pub pow: usize,
+  pub pow: u64,
 
   /// Precision (decimal places) to use
 	#[arg(short, long, default_value_t = 32)]
